frontend: add id to datagrid columns
The column filter state identifies a column by id, so if we ever want to
filter by column we need to specify its id.
It seems that if not specified it fallbacks to accessors, but these are
complicated and not the same across all datagrids.

Also changing column filters identifiers to match this.